backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
aiofiles
python-dotenv
pydantic[email]
beautifulsoup4
bcrypt
fpdf2
